zero: generate error methods for response types
